wip - moving serverless activation code to a common location
